chore: allow different aria-label
The one simplification I made in the previous commit was that the aria-label was
the same as the tooltip, but this makes it harder for e2e tests to find 'Pods'
when the tooltip is 'Pods (3)'. This change leaves the default to be the same,
but adds an ariaLabel property that allows it to be set explicitly for items
that have dynamic tooltips.

Signed-off-by: Tim deBoer <git@tdeboer.ca>