chore: reduce inkscape/sodipodi warnings in dev/test
When starting in watch mode or running tests, there is a *lot* of extraneous
console output like the following below, where sodipodi/inkscape attributes
trigger vite-plugin-svelte warnings. As an example were just shy of 400 lines
in the e2e test I was trying to debug.

By definition these attributes are not used at runtime and can be safely
removed to make these files marginally smaller and remove this output.

[vite-plugin-svelte] /src/lib/images/WindowsExitIcon.svelte:32:4 Attributes
should not contain ':' characters to prevent ambiguity with Svelte directives
30:     inkscape:window-x="0"
31:     inkscape:window-y="0"
32:     inkscape:window-maximized="1"
        ^
33:     inkscape:current-layer="g18715">
34:     <inkscape:grid type="xygrid" id="grid10569"></inkscape:grid>

Signed-off-by: Tim deBoer <git@tdeboer.ca>