fix: main nav selection
As per issue #3509, the main nav wouldn't show items as selected if they had a
space in the URL, or were child pages like settings/resources. This fixes the
first issue through URL encoding the href, and the latter by adding back the
startsWith() that was special-cased for settings before. New tests added for
both cases.

Signed-off-by: Tim deBoer <git@tdeboer.ca>