chore: link component (#3533)
We have several different styles of 'links' today:
- blue ones with underline hover (e.g. deploy to kube 'restricted')
- blue ones with light hover (e.g. deploy to kube 'Open OpenShift')
- purples ones with dark hover (e.g. deploy to kube routes, dashboard)
- purples ones with no hover (e.g. back link on detail, form pages)
- white ones with hover underline hover (e.g. dialog cancel buttons)

To make things more consistent, and obvious to the user what is
clickable, we should reuse a Link component. This PR creates a simple
component with an optional icon and either href or on:click support
for simplicity & ease of migrating from the existing a/div/spans
(which will be done via other PRs).

I tried making the Link a simpler <Button type="link">, but buttons
have padding and complexity that makes this more complex and hard to
directly reuse in either direction. IMHO it makes sense to keep Link
and Button independent but share styling; had a brief chat with
Mairin that it may make sense to change link Buttons to purple for
consistency.

Signed-off-by: Tim deBoer <git@tdeboer.ca>