chore: details & form page links
Updates the back link on details and form pages to use the Link component.
Visually the only difference is that they now change on hover.

Updated test. Since Link uses on:click we can't test for the href, but
we can go one step further instead and confirm clicking the link works.

Signed-off-by: Tim deBoer <git@tdeboer.ca>