fix: internal hrefs in link component
As per bug #3565, I managed to break the Link component so that it does not work
for internal hrefs. This changes the Link to support three options:
 - href starting with / - link uses tinro router.goto() to navigate to another page.
 - any other href - uses window.openExternal() to open in external browser.
 - on:click - used for more complicated/programmatic callbacks that need to trigger
   a function.

Added test for the first option that wasn't working before, and updates Details and
Form page tests to check the correct case.

Fixes #3565.

Signed-off-by: Tim deBoer <git@tdeboer.ca>