docs: add note python implementation ### What does this PR do?
We have two implementations, the Go implementation (compose spec binary)
or the Python implementation (containers/podman-compose).

We should indicate that this is specifically for the Python
implementation as the upcoming changes to how we use `podman compose`
will confuse the user which implementation is which.

### Screenshot/screencast of this PR

N/A

<!-- Please include a screenshot or a screencast
explaining what is doing this PR -->

### What issues does this PR fix or reference?

N/A

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

### How to test this PR?

View the site / netlify.

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>