chore: Use bootstrap color for Steps
We use the Steps Tailwind component for configuring providers. It uses Bootstrap
and defaults to using the Bootstrap primary color. Currently we override this on
each instance using a PatternFly variable (which for some historical reason is
slightly off our current palette color).

This change sets the Bootstrap primary color to the (corrected) same value as our
palette color, which means we do not need to override the value on each individual
Steps.

This is a real preference/complexity call. Ideally I'd rather use our palette
colors as CSS variables directly (but that seems over-complicated for a case like
this [1]), or we could define our own variable name and continue to use it in
each Steps. Maybe it's better to keep the override per-component so we don't
forget where it comes from? I'm proposing this change b/c it is less code and
removes one more reference to PatternFly, since we're getting close to where we
could remove it entirely. This is simplest code overall, and would work if we had
other Bootstrap components. If anyone prefers another option, please suggest.

[1] https://tailwindcss.com/docs/customizing-colors#using-css-variables

Signed-off-by: Tim deBoer <git@tdeboer.ca>