chore: remove neutral from palette
Neutral-900 was in our palette for legacy reasons, and despite a friendly
comment to not use it :), copy/paste wins and we added a new use lately. This
just replaces all uses with charcoal-900, which is almost indistinguishable
but matches other UI slightly better, and removes the color from the palette.

Signed-off-by: Tim deBoer <git@tdeboer.ca>