chore: remove blue from palette
For historical/legacy reasons we still have Tailwind blue-500 in our palette.
There are only a few references to this color and all should be replaced by
purple-500, which is in our palette and specified by the design for outlines
like these. Replacing and removing from the palette to avoid future use.

Signed-off-by: Tim deBoer <git@tdeboer.ca>