fix: simplify Windows install doc nav
I noticed when looking for the Windows install guide that the navigation
on the first page "Installing Podman Desktop on Windows" was inconsistent
with the other platforms (just "MacOS", "Linux"). Expanding it the child
nav was also very verbose/repetitive compared to elsewhere and includes
"Podman Desktop" in every title. This change make the first (index.md)
consistent with the other platforms and simplifies the rest of the pages
to match.

Signed-off-by: Tim deBoer <git@tdeboer.ca>