fix: don't prefix all podman machine names
My initial position was that we shouldn't prefix any Podman machine names, but
I've softened a bit: if someone appends to the default name and calls their
machine "podman-machine-3", then I think it's nice we pretty-print that to
"Podman Machine 3". But if I call my machine "My Machine", and that's the name
I see on the command line, then I still believe we should use "My Machine" and
not "Podman Machine My Machine".

I have been using this change for a few days, and I prefer it. The only place
I've found where there isn't the obvious context of My Machine being a Podman
machine is in the system tray, and IMHO that's not a problem because I named
it and recognize what it is in that context.

Fixes #3424.

Signed-off-by: Tim deBoer <git@tdeboer.ca>