fix: initial action state in details pages
While testing issue #3529 I noticed that when you're in the pod or container
details page, all actions change the state to STARTING regardless of what the
action says the state should be changed to. If you were actually deleting the
object this means there's often a little blip where it looks like it's trying
to start before the state is corrected.

Fixed, now matches the correct/identical code that's in the corresponding
*List pages.

Signed-off-by: Tim deBoer <git@tdeboer.ca>