chore: deleting status icon
Currently there is no indication in the status icon when something is being deleted.
We're adding text to the age column (#4056), but the design preference is that this
is re-enforced by the status icon changing into a spinner.

This adds the DELETING status, using a correctly sized Spinner component with no
border. Added a test, and some missing tests for other states.

Signed-off-by: Tim deBoer <git@tdeboer.ca>