fix: run image error spacing
Fixes error message over top of the field and label as shown in issue #3628.
You can notice the Entrypoint field in the issue is also very close to the
following Command label - both fields did not have the pt-4 padding that every
other field in this form has. Adding this back leaves a blank line for the
error, so I added an if statement as we do in many other uses of ErrorMessage.

Fixes #3628.

Signed-off-by: Tim deBoer <git@tdeboer.ca>