fix: solid pod icon
Fixes the 'large pod icon regression' when adopting the Button component and using
PodIcon. Adds a SolidPodIcon with a more natural default size (1em) and removes
the failing custom check in Button.

We may have to do the same with ContainerIcon, ImageIcon, and VolumeIcon at some
point, but not doing it now since these components are not used. Likewise we may
want to change the default size of all of them to 1em, but that's an independent
change.

Fixes #4090.

Signed-off-by: Tim deBoer <git@tdeboer.ca>