fix: improve featured extension breakpoints
The responsive breakpoints for featured extensions were set too small, so in the
Settings > Extensions page it would scale to where the icons were not visible.

Responsive breakpoints are set based on window width, so we have a bit of a
dilemma since this component is shown in both the Dashboard (with minimal left
navbar and infinite width) and Settings > Extensions (bigger left navbar and
width limit). I tried to balance these, so the columns get a little big in the
Dashboard and don't go past three columns, but don't cause noticeable scaling
issues in Settings.

Also moved padding from the bottom of this component to the Dashboard where it
was missing.

Fixes #3460.

Signed-off-by: Tim deBoer <git@tdeboer.ca>