fix: remove suggested registries on deactivation
The suggested registries extension adds registries but never adds the disposable
subscription, so registries stay around even if you disable the extension. This
just adds the subscription.

There's a good argument to be made that suggested registries should be registered
more like commands or menu contributions, and since they'd be handled once by the
extension loader this bug could never happen again. That would require a breaking
API change and a much bigger PR though, so not sure it's worth it; either way I
think it's best for this change to go ahead and fix the immediate bug.

Fixes #3909.

Signed-off-by: Tim deBoer <git@tdeboer.ca>