chore: show onboarding for unknown status providers
### What does this PR do?

Some providers are using 'unknown' such as compose and kind in order for
them NOT to appear in the dashboard.

This allows the new compose onboarding to be able to still show
enablement / onboarding process.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast
explaining what is doing this PR -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/4207

### How to test this PR?

<!-- Please explain steps to reproduce -->

Run the tests, or patch the PR against
https://github.com/containers/podman-desktop/pull/3953 it will not show
the "Setup..." button.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>