fix: tooltip z order
I have no idea how we didn't notice this before now, but since we fixed some
other z-ordering issues in May, the tooltip has put a higher z-order not just
on the tooltip itself but _also on the component that is getting the tooltip_.

This means that if you have two objects that both have tooltips close
enough to each other, depending on the positioning of the tooltip one
of the objects will always be over top of the other's tooltip. In my case I'm
adding tooltips to two Donuts and the tooltip remains behind the following
Donut.

The fix just moves the z-order to only be on the tooltip div.

Signed-off-by: Tim deBoer <git@tdeboer.ca>