chore: donut design and fixes
Updates the Donut component to match the design better, and fixes a few minor bugs:

- Grey circle and arc are thinner (closer to the design).
- Switched to using colors from the palette.
- Since x,y=radius and starting angle=0 always I simplified the function a bit.
- Made the viewbox square, reduced padding, and made font sizes relative to reduce
  scaling issues.
- Added end limits (before if the client incorrectly requested 125% the arc would
  fail, now it looks like 100%).
- Colors weren't changing if the live value crossed thresholds (e.g. green to red),
  made it dynamic.
- Added tooltip.

There are still several issues compared to the design, e.g.:
- The titles doesn't match.
- Unit font should be smaller.
- Values should be rounded more (e.g. should be 99Gb not 99.25Gb).

Fixing the first is likely better done in the calling code, e.g.: changing the
property names or some custom code to override 'disk space' with 'disk'. Unless
we want the Donut to be really clever the second likely should be fixed in the
caller too. Either way, it felt like there are more than enough changes in one
PR and better to stop here for now.

Signed-off-by: Tim deBoer <git@tdeboer.ca>