chore: use donuts for container stats
Switches the container statistics from bars to use the Donut component.
This is more consistent, will match what we do in Settings > Resources,
visually nice, and takes up far less horizontal space vs today (which
is necessary for #4226 - details page consistency with design and other
forms).

Minor benefit: the colors are now from the palette.

The donuts are taller than the previous component, but still visually
look a little small. To me this is ok and still legible, and I can't go
any bigger and solve #4226.

Two minor fixes for Donut: move the tooltip below instead of right (since
container stats are shown near the right edge of the window) and do not
show the value until it is set. This avoids showing 'undefined' before the
container stats are available.

Partial fix for #4226.

Signed-off-by: Tim deBoer <git@tdeboer.ca>