feat: provider creation form
Creating an instance of a Podman Machine or Kubernetes provider currently has
a UI that doesn't quite match other places. Switching it to be a FormPage
aligns the look and feel with the rest of our UI (header, breadcrumbs, etc.)
and fixes things that we've already solved once in all the form pages,
e.g. clicking Esc to close (#2853).

Since this page was done a while ago there was some minor restructuring to
make this happen:
- The Resources page has to be marked as a details navigation hint so that
  the backlink/close button goes back to the correct page.
- The code to find the provider display name and image (header info) is moved
  up from the connection creation page to the provider rendering form and
  onboardig.
- This makes the provider image 'always hidden' on the component, so the
  property could be removed.

Fixes #2853.

Signed-off-by: Tim deBoer <git@tdeboer.ca>