fix: add error to analyzed extension
Adds an error property to analyzed extensions, and analyzeExtension() sets
it whenever there is a problem loading a folder instead of returning
undefined. Updated callers to check for error instead of undefined, and
send an error when trying to install an extension that has an error.
Tests updated to confirm an error is being created.

Signed-off-by: Tim deBoer <git@tdeboer.ca>