fix: details page consistency (#4370)
* fix: details page consistency

Moves the close button and actions of the details page to be consistent with
form pages. It also makes the subtitle font a little smaller, and adds a
missing shadow-pageheader that was in the design.

The overall page structure & layout is much closer to form pages after this.
I didn't go full shared-component because there are still a few differences
between the two, but this is a step towards a common page component or
breadcrumb + close header.

Shouldn't really be released before PR #4227. It's not awful, but one of the
reasons for switching to donuts is that it takes up less horizontal space and
won't go over the tab titles.

Fixes #4226.

Signed-off-by: Tim deBoer <git@tdeboer.ca>