fix: log kind cluster creation failure with usage
Minimal change to send one event when we create a kind cluster instead of 1 event + different event on failure.

Added the duration while I was at it, but nothing else. There is likely more work to do here to classify failures or add more information to tell what is different about failure cases, but this will solve one core problem that it isn't easy to correlate failure with the usage.

Fixes part of #4214.

Signed-off-by: Tim deBoer <git@tdeboer.ca>