fix: allow editing of build containerfile
Changes the build image form from a readonly text field that you click on to
browser, to a more common editable text field with a Browse button. Likewise
the build directory was hidden until you entered a containerfile before, now
it is always visible and uses the same text field + browse button.

This change makes the form use a more common pattern, and allows e2e tests
to more easily enter text in either field.

The drawback to this approach is that there is no more validation than
before - the build button enables whenever there is text in both fields.
This happens automatically when you use the Browse button like before,
but if you edit or copy/paste and have a typo you won't find out until you
click Build. We could add more validation in the future, but IMHO the
better usability/form layout is enough of a benefit to make this change in
the meantime.

Signed-off-by: Tim deBoer <git@tdeboer.ca>