refactor: remove compose from the status bar
### What does this PR do?

Removes compose from the status bar since we are not implementing
onboarding.

This does NOT remove the excess code from `compose-extension.ts` yet.
This will be need to be cleaned up in the future.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast
explaining what is doing this PR -->

N/A, should not appear anymore

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/3809

### How to test this PR?

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>