fix: add autoscroll to summary pages
None of the summary pages had overflow, so although most were small all would get cut-off if the UI is small enough or (e.g.) a container id is long enough.

Also changes my- to mt- in pods and volumes since this was causing a large bottom gap instead of just between the sections.

Signed-off-by: Tim deBoer <git@tdeboer.ca>