fix: remove prev/next bar
Users don't read documentation like a book anymore, and when there is an
obvious next step we're already putting this in a Next Steps section. To
avoid the duplication and spending any time trying to make the prev/next
bar consistent, we should just remove this bar entirely.

As per [1] the Docusaurus feature to allow you to disable prev/next may
be coming in a future release, so the only current options are to add
  pagination_next: null
  pagination_prev: null
in every page, or use custom css to hide the bar. This does the latter.

[1] https://github.com/facebook/docusaurus/discussions/6525

Signed-off-by: Tim deBoer <git@tdeboer.ca>