feat: reorganize doc navigation by provider
Switches to give every provider (Podman, Lima, Kind, Minikube, OpenShift) its own root folder in the doc navigation (and the child '/installing' is normally where installation instructions are).

Why? Currently the topics for several of these are spread out between multiple sections, e.g. there is Lima information in onboarding for containers, onboarding for Kubernetes, and Kubernetes; installing Podman is under Onboarding for containers but setting the default connection is in Working with containers. As a user I know if I'm using provider X and if I'm looking for information on using X within Podman Desktop, vs how to use Podman Desktop.

After this move, there was only one page in each of the 'onboarding' sections - the onboarding containers overview and working with an existing Kubernetes. These two pages were moved into the respective /containers and /kubernetes folders so that the section could be removed. /working-with-containers was reduced to /containers to match all the other folders.

As much as possible (except for links) I've tried not to edit the content of any page. In a couple cases I had to merge pages that were near duplicates, and likewise there are a couple places where we didn't have a single landing page before (e.g. OpenShift) where I added an initial page.

Added redirects in the docusaurus config, and as much as possible updated links in existing files to point to the new locations.

Signed-off-by: Tim deBoer <git@tdeboer.ca>