fix: reduce website footer (#4546)
* fix: reduce website footer

Ok, here's my pass at reducing the vertical height of the footer. Long term I think
we can and should do more, this is just a quick initial cleanup pass:
- Left: removed three links (so both columns are down to 5). These are a
  judgement call, e.g. I felt like 'Docker migration' is a bit of marketing we'd
  want to leave for now.
- Right: remove Blog (it's already at the top of the page).
- Right: remove planning and roadmap, both not really useful link and accessible
  from the Github link above.
- Put copyright and license on one line.

Overall, this reduces the footer height from 11 lines of text to 7.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* fix: chat, sprint links

Reduced four chat links to just Discord and a link to Communication section in the readme. Added back in a link to the sprint planning board.

Removed Compose from the doc links to even things out again.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>