nit: remove padding from link
### What does this PR do?

There is a little bit of padding in the link which makes it difficult to
implement columns / tables with lining up the imported link to other
elements.

It also fixes the very subtle spacing in parapgrahs where link is being
used.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast
explaining what is doing this PR -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Found it while implementing: https://github.com/containers/podman-desktop/pull/4560

### How to test this PR?

<!-- Please explain steps to reproduce -->

Go to "Deploy To Kube" page and see the link (it should be more inline
with the text)

Signed-off-by: Charlie Drage <charlie@charliedrage.com>