fix: reduce size of provider cards on the dashboard
The current provider cards on the dashboard are massive, and it restricts us from
adding other things to the dashboard, e.g. notifications or tutorials. There is a
longer term redesign of these cards in the works, but in the meantime we can easily
start by reducing their size.

Made the following changes:
- Reduce logo size from max-h-24 to max-h-12.
- mt-10 is excessive, reduce all to margins to mt-5.
- provider.updateInfo?.version can be undefined when there is no update available, so adding a check removes the section.
- ProviderWarnings section was always there. Change it to be like the other sections and only appear when there is content.
- Remove extraneous <div> in ProviderInstalled.

Fixes part of #4395.

Signed-off-by: Tim deBoer <git@tdeboer.ca>