feat: add doc for telemetry and usage data
The telemetry investigation in #3788 led to a couple issues being opened to improve
the events we're sending, but one of the comments was that we don't really have this
written down anywhere.

Some other projects contain a USAGE_DATA.md file, so this is my attempt to both
write down what we collect and why for users, but also start some coding notes for
committers. Feedback appreciated.

Fixes #3788.

Signed-off-by: Tim deBoer <git@tdeboer.ca>