chore: move container and image list
I know these were probably created first and they are special, but it still throws
me off when ContainerList and ImagesList are not in the folders like the equivalent
PodsList, VolumesList, etc. Just moving these to follow the pattern and reduce
files in the common folder.

Also fixing the typo in ImagestList.spec.ts.

Signed-off-by: Tim deBoer <git@tdeboer.ca>