fix: light scrollbars and line height
After the PatternFly removal I noticed that the scrollbars on Mac are in light-mode.
PF-dark was setting the color-scheme before, setting it again fixes the problem and
means we can also remove the explicit text-white. (maybe some of the changes in
#4803 too?)

PF was also setting the line height to 0.9. I'm not tied to it, but it makes buttons,
multi-line text, and preferences nav items all less tall. Maybe I'm just used to it,
but I prefer a smaller window and 0.95 gets most of the way there & uses less space.

Signed-off-by: Tim deBoer <git@tdeboer.ca>