fix: merged compose deploy to kube page in UI
When used on a compose app, the Deploy to Kube form incorrectly shows the
compose details at the bottom of the page. You can also see it in the breadcrumb
that says 'Container > Compose Details' when it should say 'Container > Deploy
to Kubernetes'.

The problem is the route path of these two pages can overlap:
 - /compose/deploy-to-kube/:composeGroupName/:engineId/*
 - /compose/:name/:engineId/*
i.e. deploy to kube path also resolves to compose details page with name
'deploy-to-kube' and engineId of the compose group name.

Simple fix to just rename the path so they don't overlap.

Signed-off-by: Tim deBoer <git@tdeboer.ca>