feat: table component
This is a clean branch of the table component after previous review comments
in PR #4545.

Each page passes an array of objects and Columns to the Table component, which
is responsible for creating the basic layout and calling back to the Column
renderer to render each cell in the table. The Table uses grid layout and
provides enough abstraction that we can implement additional features later.
(e.g. optional columns, column reordering, or resizing. grouping. etc)

The Table provides sorting & sort indicators, and tests are included.
Support for 'object containers/grouping' in order to support Container list
will be done separately since this is already a big PR.

Fixes #4365.

Signed-off-by: Tim deBoer <git@tdeboer.ca>