feat: volumes using table component
Switches the Volume page to the new table component. Columns are extracted
as expected, and I've pulled out the engine into an Environment column to
match the Pods page.

I made all the columns sortable. The raw size had to be made visible from
VolumeInfoUI because the regular column is humanized and not sortable.
Status and size are 'reverse sorted' because you typically want to see
used volumes and bigger things first.

Based on the Table PR #4545. Ignore everything in /table/* and once that
has merged I'll rebase.

Signed-off-by: Tim deBoer <git@tdeboer.ca>