feat: images using table component
Switches the Images page to the new table component. Columns are extracted
as expected, and I've pulled out the engine into an Environment column to
match the Pods page.

I made all the columns sortable. The raw size had to be made visible from
ImageInfoUI because the regular column is humanized and not sortable.
Status and size are 'reverse sorted' because you typically want to see
used images and bigger things first.

Signed-off-by: Tim deBoer <git@tdeboer.ca>