feat: default table sorting
Adds support for a 'default sort column' so that you can tell if the data was
initially sorted by a particular column (and don't try to sort that column
and have nothing change).

Dims the sort indicator on non-active columns so that you can still tell which
columns are sortable, but they don't stand out as much.

Also includes a minor '&&' simplification suggested in previous PR review.

Fixes #4838.

Signed-off-by: Tim deBoer <git@tdeboer.ca>