feat: deployment store
Follows the same pattern as pods, images, volumes, etc to define a deployment store
that's reactive to the expected events, a UI type, and a util class that maps to the UI
type.

Re: index.ts - I copied from Luca, who likely copied from elsewhere, and we both missed
updating one type. :)

We will definitely need additional properties mapped from V1Deployment to DeploymentUI
when we add the UI, this PR is purely to get the store, UI object, and basic scaffolding
in place.

Signed-off-by: Tim deBoer <git@tdeboer.ca>