feat: table sorting II
Fixes the sort issues raised in issue #4881:

- Fixed sort direction: ascending sort should be 'up', descending 'down'.
- Added a new 'initialSortDir' for Columns like Size to specify that they
  should be sorted in descending order by default. This could have been a
  boolean flag, but seeing "initialSortDir: 'descending'" in Column
  definitions seems more natural and obvious to me.

This also fixes the sorting gap raised in #4860: constant sorting by any
column.

Prior to this change sorting only happens once, when the user clicks on
the column header. If the data changes it becomes unsorted again, and the
column header is then wrong/out of sync. e.g. if you're in the Containers
view, sort by Status, and then start a new container, data will revert
back to the initial order and the new container will appear where it
would have, even though the Status column still says it is sorted.

Now, sorting is also applied _whenever the data changes_. This means that
if you sort by Status and start a container it would move to the top, or
if you've sorted Age by descending it would move to the bottom. i.e. the
sorting is always applied. This also means that any view can set the Table
defaultSortColumn to any arbitrary column.

Fixes #4881.

Signed-off-by: Tim deBoer <git@tdeboer.ca>