feat: service store
Follows the same pattern as pods, images, volumes, etc to define a service store that's reactive to the expected events, a UI type, and a util class that maps to the UI type.

We will likely need additional properties mapped from V1Service to ServiceUI when we add the UI, this PR is to get the store, UI object, and basic scaffolding in place.

Signed-off-by: Tim deBoer <git@tdeboer.ca>