feat: deployments watch
Adds a watch to deployments, identical to what we're doing for pods. This is
not a long-term solution for #4974, but for now it is simple and responds to
changes in deployments, allowing PR #4931 to proceed.

We clearly need to do #4974 at some point, but I think we should be able to
do this for ingress/routes/services without issue? I could do all of these
in this PR.

Signed-off-by: Tim deBoer <git@tdeboer.ca>