feat: deployment empty screen, nav hiding, tests
Hides deployment icon from the navigation when there are no Kubernetes
contexts.

Adds empty screens: no deployment and when search is filtered. There is no need
for a 'no provider' screen like other list pages because Deployments is hidden
from the UI in that case, and because for deployments there might not actually
be a Kubernetes provider: you could just have a remote context you added
externally.

Adds missing tests for empty screens and basic UI.

Signed-off-by: Tim deBoer <git@tdeboer.ca>