feat: provider cards
Creates a ProviderCard component and makes each of the Provider* components use
it in order to get some common layout and code between the different states.

With the card I tried to make it more like the example in #4395 and our other
cards (e.g. Resources, Kubernetes contexts, or CLI), creating two columns with
the first having the common bits: logo at the top-left corner with the name and
version, with status below it, and the right containing everything else that is
specific to the state.

The existing status dots + text components don't quite work for providers so I
created a new one. Need to review this, overall componentization, and create
tests for ProviderCard before raising for official review.

Signed-off-by: Tim deBoer <git@tdeboer.ca>