fix: revert styling of disabled buttons
When tab button type was added for the All/Stopped/Started containers, it
removed the rounded/text styling from disabled buttons. As a result, there
was a visible difference between disabled/in progress buttons and regular
buttons.

This adds these two styles to all non-tab buttons, regardless of whether
they are regular/disabled/in-progress.

Fixes #4882.

Signed-off-by: Tim deBoer <git@tdeboer.ca>