feat: extension icons
After seeing Luca's PR #5089 I felt like I had to do this. :) Uses the
extension's icon in the Settings > Extensions page, using an almost
identical component to ProviderLogo. If the extension has no icon, the
puzzle piece is still used.

I've put the component in the /preferences folder since that's the only
place it is being used for now.

The puzzle pieces used purple vs gray to show whether the extension is
running, I've switched to using a brightness filter instead, since it
is consistent between the two. Tests included.

Fixes #5100.

Signed-off-by: Tim deBoer <git@tdeboer.ca>