feat: icon image component
Creates a new IconImage component (couldn't think of a better name) that
can display our normal icon/logo pattern of "string | { light: string;
dark: string } | undefined" as an image, with the default slot as a
fallback.

The logic for which image to use could have been directly in the
component, but in order to start sharing code with Appearance.svelte (and
in case we have other places that will need to know which theme we're
using) I extracted the basic function to appearance-utils.js.

This component would allows us to basically delete components like
ProviderLogo and ExtensionIcon and replace with single lines like:
<IconImage image="{provider?.images?.logo}" class=".." alt="{provider.name}"/>

Signed-off-by: Tim deBoer <git@tdeboer.ca>