chore: simplify pod actions & callbacks
It has been bugging me for a while that PodActions has a callback when an
action affects a pod, and both PodList and PodDetails have essentially the
same response of updating the pod UI object, and then PodList needs to
trigger the UI to update. It also passes a podInfoUI to functions when
PodActions already has the pod as a property.

When trying to extract columns to use the Table component this gets
uglier because there is an intermediate component that needs to pass the
callbacks through as well, so I figure it's a good time to simplify this
a bit first.

This change just makes the PodAction update the pod UI object itself, and
then uses a standard dispatch event to notify parents that something
changed. Both PodList and PodDetails just need an on:update to trigger a
Svelte update. Simpler/less code all around and less chance of PodList
and PodDetails having different behaviour.

Note that there's one thing that still bothers me but is not fixed in this
change: PodList and PodDetails use their own copies of PodInfoUI objects,
so if you click Start in one view and then switch quickly, the pod is
still stopped for a couple seconds in the other view. That is a separate
problem but at least this change might make it marginally easier to fix
in the future.

Signed-off-by: Tim deBoer <git@tdeboer.ca>