chore: adjust details heading and summary post-pf
While testing another PR I noticed that the DetailsPage header was taking
up a lot of vertical space. The Summary tabs are especially poor and the
rows are really spaced out. I checked and sure enough, we were adding
extra padding to compensate for PatternFly's narrow line-height, so when
we removed PF these gaps became noticeable.

This PR removes all unnecessary padding from each of the Summary pages,
and makes minor changes to padding in Tab and DetailsPage to return to
the original design.

While I was there I added an if around the tabs, so if we ever have a
details page without tabs there won't be an empty div adding padding
and a line.

Signed-off-by: Tim deBoer <git@tdeboer.ca>