feat: reading deployments, ingress, routes, and services
When we do DetailPages for the four Kubernetes objects that we're
building nav pages for, we'll need the ability to read each of the
objects the same way we do for pods today.

These are fairly boilerplate, so I'm just doing them all at once.
Extracted createTestClient() since we're calling that a lot.

Signed-off-by: Tim deBoer <git@tdeboer.ca>