fix: allow table columns to specify overflow
With the previous table (<1.6), column widths were variable and this
caused problems:
- sometimes a long image name would cause the tables to require
  horizontal scrolling
- sometimes a change in one column would cause other columns to shift
  E.g. starting a container can cause the start/stop button to jump
  right, and stopping the container the button jumps left, so you
  might click the wrong button (e.g. delete).

With the new grid-based table component, column widths are fixed so
content doesn't affect table width. This solves the previous issues,
but it means that simple columns with long content are more likely to
have overflow. This was 'fixed' in #4841 by setting a maximum column
width and overflow-hidden to limit rendering to within the cell, but
now things like popup menus can't go outside of the cells.

I've struggled deciding what the correct default behaviour should be,
but I still think #4841 is correct and means the average renderer can
stay simple. This change adds a new 'overflow' property to allow
renderers that are more interesting to control their own behaviour
and render 'outside' of their cell.

Property added to both image and volume action columns.

Fixes #5220.

Signed-off-by: Tim deBoer <git@tdeboer.ca>