fix: improve default contribution action icon
We've had several people get confused with actions contributed by
extensions that show up as an ellipsis icon in popup menus and Detail
pages. The correct/full fix will require us to allow commands to provide
their own icons, and to do this for all of the current extensions. This
is not that fix. :)

The other two things we should do is switch the default icon so that it
doesn't use the same icon as our popup menus, and to pass the detailed
property through the ContributionActions component so that these icons
show up on buttons on the Details pages like the other actions. This is
that part of the fix!

I went through FA trying to find a more suitable icon and this is the
best I could come up with. I preferred it to icons like the puzzle
piece and I think it still gives some indication that the action comes
from an extension.

First step toward fixing #4763.

Signed-off-by: Tim deBoer <git@tdeboer.ca>